==> analysis/1_processing.Rmd <==
---
title: "Visual Search and Pre-crastination"
author: "Alasdair D.F. Clarke"
date: '2022-05-30'
output:
  tufte::tufte_html:
    number_sections: true
    fig_height: 3
    link-citations: yes  
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(digits = 3)
set.seed(2022)
```

```{r, message=FALSE}
# attach packages
#library(rethinking)
library(tidyverse)
library(patchwork)
library(binom)
```

```{r}
# set ggplot2 options
theme_set(ggthemes::theme_tufte() + 
            theme(plot.background = element_rect(fill = "#fffff8", 
                                                 colour = "#fffff8")))

our_cols <- c("#1b9e77", "#7570b3")
options(ggplot2.discrete.fill = our_cols ,
        ggplot2.discrete.colour = our_cols)
```

# Overview of Visual Search Task

Import data
```{r}
# read in trial data (reponse times and accuracy)
dat_trl <- read_csv("../data/shls_rt_acc.csv", col_types = "cddcdd")

# read in fixation data (only first 20 fixations... see ascii files for full raw data)
dat_fix <- read_csv("../data/processed_fix_data.csv", col_types = "ddddddf")
```

Let's count participants:

```{r}
length(unique(dat_trl$person))
length(unique(dat_fix$person))
```

## Reaction Time and Accuracy

```{r}
dat_trl %>%
  group_by(site, person, targ_side_rel) %>%
  summarise(median_rt = median(rt),
            accuracy = mean(accuracy), 
            .groups = "drop") %>%
  mutate(targ_side_rel = fct_relevel(targ_side_rel, "absent")) -> dat_rt_acc

# plot
ggplot(dat_rt_acc, aes(x = median_rt, y = accuracy, colour = site)) + 
  geom_point(alpha = 0.5) +
  scale_x_log10() + 
  geom_hline(yintercept = 0.75, linetype = 2) + 
  facet_wrap(~targ_side_rel)
```

Our pre-registered exclusion rule was to remove anybody achieving less than 75% accuracy in the *absent* and *easy* conditions.

```{r}
dat_rt_acc %>% 
  filter(targ_side_rel != "hard", accuracy < 0.75) %>%
  select(person) -> to_remove

peeps_to_remove <- unique(to_remove$person)

dat_rt_acc %>% 
  filter(!(person %in% peeps_to_remove)) -> dat_rt_acc

# go back and remove from the larger dataset of fixations
dat_fix %>% filter(!(person %in% peeps_to_remove)) -> dat_fix

rm(to_remove)
```

This gives `r length(peeps_to_remove)` participants to remove, leaving us with `r length(unique(dat_rt_acc$person))` participants left (for the visual search part of the experiment).

THIS NEEDS DONE LATER ONCE WE HAVE THE PRECRASTINATION DATA MERGED WITH IT

Now let's do a final plot of accuracy and reaction time:

```{r}
ggplot(dat_rt_acc, aes(x = targ_side_rel, y = accuracy)) + 
         geom_boxplot() -> plt_acc

ggplot(dat_rt_acc, aes(x = targ_side_rel, y = median_rt)) + 
         geom_boxplot() +
  scale_y_log10() -> plt_rt

plt_acc + plt_rt

rm(plt_acc, plt_rt)
```

## Search Strategy

Now to compute our search strategy metric

```{r}
# classify every fixation as homo (right), central, or hetro (left)
centralWidth = 64 #change to 1 visual degree
dat_fix$side <- 'central'
dat_fix$x <- dat_fix$x_flip

dat_fix$side[which(dat_fix$x < (0 - centralWidth/2))] = "hard"
dat_fix$side[which(dat_fix$x > (0 + centralWidth/2))] = "easy"
dat_fix$side = as.factor(dat_fix$side)

dat_fix %>% left_join(dat_trl, by = c("person", "trial", "targ_side_rel")) %>%
  filter(side != "central",
              n < 6, n > 1, 
              targ_side_rel == "absent",
              accuracy == 1) %>% 
  group_by(person, n) %>% 
  summarise(
    prop_hard = mean(side == "hard"), 
    n_trials = length(trial),
    lower = binom.confint(prop_hard*n_trials, n_trials, method = 'wilson')$lower,
    upper = binom.confint(prop_hard*n_trials, n_trials, method = 'wilson')$upper,
    .groups = "drop") -> dat_strat

# also calculate aggregate overall strat and put with rt and acc
dat_strat %>% group_by(person) %>%
  summarise(prop_hard = mean(prop_hard), .groups = "drop") %>%
full_join(dat_rt_acc, by = c("person")) %>%
  mutate(targ_side_rel = fct_relevel(targ_side_rel, "easy","hard")) -> dat_shls

write_csv(dat_shls, "scratch/dat_rt_acc_strat.csv")
```

# Pre-crastination

## Do people pre-crastinate?

```{r}
dat <- read_csv("../data/pre_crastinate.csv", 
                col_types = "iif") 

summary(dat)
```

We have data from `r nrow(dat)/6` people and no missing data. 

First, report number of people who *double pre-crastinate*:

```{r}
dat %>% filter(choice == "double_pre_cra") %>% 
  group_by(person, site) %>%
  summarise(n_trials_precrastinated = n(),
            .groups = "drop") %>%
  group_by(site, n_trials_precrastinated) %>%
  summarise(num_people = n(), 
            .groups = "drop") -> dpc

knitr::kable(dpc)
```

We find that `r sum(dpc$num_people)` out of `r length(unique(dat$person))` participants *double pre-crastinated* at least once. Out of these, `r filter(dpc, n_trials_precrastinated==6)$num_people` did so on all six trials.

As this is another form of pre-crasination, we will not distinguish between these participants and those who pre-crastinate normally. 

```{r}
rm(dpc)
```

```{r}
dat %>% 
  group_by(person, site) %>%
  summarise(
    sensible = sum(choice == "sensible"),
    pre_crastinate = sum(choice != "sensible"),
    .groups = "drop") %>%
  mutate(
    classification = case_when(
      sensible > 4 ~ "no",
      sensible < 3 ~ "yes",
      TRUE ~ "intermediate"),
    classification = as_factor(classification),
    classification = fct_relevel(classification, 
                                 "no", "intermediate", "yes")) %>%
  select(person, site,
         n_pc = pre_crastinate, 
         pre_crastinate = "classification") -> dat_pc

```

# Personality

Read in personality data and compute overall scores.

```{r, warning=FALSE, message=FALSE}
dat_p <- read_csv("../data/personality_q.csv") 
```

We have data from `r length(unique(dat_p$person))` people. 

```{r}

# creating list of questions
procras_qs <- names(dat_p)[str_detect(names(dat_p), "ProScale")]
bfi_qs <-     names(dat_p)[str_detect(names(dat_p), "BFI")]
bis_qs <-     names(dat_p)[str_detect(names(dat_p), "BIS")]

# variables to reverse code
rev_code <- c("BFI2_03", "BFI2_04", "BFI2_05", "BFI2_08", "BFI2_09", "BFI2_11", 
              "BFI2_12", "BFI2_16", "BFI2_17", "BFI2_22", "BFI2_23", "BFI2_24", 
              "BFI2_25", "BFI2_26", "BFI2_28", "BFI2_29", "BFI2_30", "BFI2_31", 
              "BFI2_36", "BFI2_37", "BFI2_42", "BFI2_44", "BFI2_45", "BFI2_47", 
              "BFI2_48", "BFI2_49", "BFI2_50", "BFI2_51", "BFI2_55", "BFI2_58", 
              "ProScale_03", "ProScale_04", "ProScale_06", "ProScale_08", "ProScale_11", 
              "ProScale_13", "ProScale_14", "ProScale_15", "ProScale_18", "ProScale_20", 
              "BIS11_01", "BIS11_07", "BIS11_08", "BIS11_09", "BIS11_10", "BIS11_12", 
              "BIS11_13", "BIS11_15", "BIS11_20", "BIS11_29", "BIS11_30")

for (var in rev_code) {
  dat_p <- mutate(dat_p, "{var}r" := max(!!as.name(var), na.rm = T) - !!as.name(var) + 1) 
}
rm(var)

# create Big 5 measures
# 
# Extraversion: 1, 6, 11R, 16R, 21, 26R, 31R, 36R, 41, 46, 51R, 56
# Agreeableness: 2, 7, 12R, 17R, 22R, 27, 32, 37R, 42R, 47R, 52, 57
# Conscientiousness: 3R, 8R, 13, 18, 23R, 28R, 33, 38, 43, 48R, 53, 58R
# Negative Emotionality: 4R, 9R, 14, 19, 24R, 29R, 34, 39, 44R, 49R, 54, 59
# Open-Mindedness: 5R, 10, 15, 20, 25R, 30R, 35, 40, 45R, 50R, 55R, 60

dat_p %>% select(person, starts_with("BFI2_")) %>%
  pivot_longer(-person, names_to = "item", values_to = "resp") %>%
  separate(item, into = c("BFI", "item")) %>%
  select(-BFI) %>%
  pivot_wider(person, names_from = "item", values_from = "resp") %>%
  mutate(extraversion = 
           rowMeans(select(., '01', '06', '11r', '21', '26r', '31r', '36r', '41', '46', '51r', '56')),
         agreeableness = 
           rowMeans(select(., '02', '07', '12r', '17r', '22r', '27', '32', '37r', '42r', '47r', '52', '57')),
         conscientiousness = 
           rowMeans(select(., '03r', '08r', '13', '18', '23r', '28r', '33', '38', '43', '48r', '53', '58r')),
         negative_emo = 
           rowMeans(select(., '04r', '09r', '14', '19', '24r', '29r', '34', '39', '44r', '49r', '54', '59')),
         open_mindedness = 
           rowMeans(select(., '05r', '10', '15', '20', '25r', '30r', '35', '40', '45r', '50r', '55r', '60'))) %>%
  select(person = "person", extraversion, agreeableness, conscientiousness, negative_emo, open_mindedness) -> d_BFI

# create procrastination variable
procras_qs_rev <- paste0(procras_qs[which(procras_qs %in% rev_code)], "r")
procras_qs_fwd <- procras_qs[which(!(procras_qs %in% rev_code))]
procras_qs <- c(procras_qs_fwd, procras_qs_rev)
rm(procras_qs_fwd, procras_qs_rev)

dat_p %>% pivot_longer(procras_qs, names_to = "ProscaleID", values_to = "score") %>%
  group_by(person) %>%
  summarise(procrastination =  mean(score)) %>%
  rename(person = "person") %>%
  full_join(d_BFI, by = "person") -> dat_p2

rm(d_BFI)

dat_p %>% select(person, starts_with("BIS11_")) %>%
  pivot_longer(-person, names_to = "item", values_to = "resp") %>%
  separate(item, into = c("BIS", "item")) %>%
  select(-BIS) %>%
  pivot_wider(id_cols = person, names_from = "item", values_from = "resp") %>%
  mutate(BIStotal = 
           rowMeans(select(., '01r', '02', '03', '04', '05', '06', '07r', '08r', '09r', '10r', '11', '12r', '13r', '14', '15r', '16', '17', '18', '19', '20r', '21', '22', '23', '24', '25', '26', '27', '28', '29r', '30r')),
         attentional_imp = 
           rowMeans(select(., '05', '09r', '11', '20r', '28', '06', '24', '26')),
         motor_imp = 
           rowMeans(select(., '02', '03', '04', '17', '19', '22', '25', '16', '21', '23', '30r')),
         nonplan_imp = 
           rowMeans(select(., '01r', '07r', '08r', '12r', '13r', '14', '10r', '15r', '18', '27', '29r'))) %>%
  select(person = "person", BIStotal, attentional_imp, motor_imp, nonplan_imp)  %>%
  full_join(dat_p2, by = "person") -> dat_p2

dat <- full_join(dat_p2, dat_pc) %>% 
  full_join(dat_p %>% select(person, bead_colour, n_beads, Vsearch_YesNo, Bucket_YesNo))

write_csv(dat, "scratch/prcrastination_and_personality.csv")

```
